Clean up header inclusion.
2008-12-01  Jim Huang  <jserv.tw@gmail.com>

        * include/global.h,
        include/internal/chewing-private.h,
        include/internal/chewing-utf8-util.h,
        src/hanyupinyin.c,
        src/chewingio.c,
        src/chewingutil.c,
        src/hash.c: Remove the inclusion of header <stdlib.h> and <stdio.h> in
        Chewing API headers.  Place them in implementation C source files
        instead.


git-svn-id: https://svn.csie.net/chewing/libchewing/trunk@1014 ac1be623-90ea-0310-9410-ba68b2efa777

+++ src/hash.c
@@ -17,6 +17,8 @@
 /* ISO C99 Standard: 7.10/5.2.4.2.1 Sizes of integer types */
 #include <limits.h>
 #include <assert.h>
+#include <stdlib.h>
+#include <stdio.h>
 
 #include "chewing-utf8-util.h"
 #include "hash-private.h"
@@ -520,7 +522,7 @@ int InitHash( const char *path )
 	} else {
 		sprintf( hashfilename, "%s" PLAT_SEPARATOR "%s", path, HASH_FILE );
 	}
-	memset( hashtable, NULL, HASH_TABLE_SIZE );
+	memset( hashtable, (int) NULL, HASH_TABLE_SIZE );
 
 open_hash_file:
 	dump = _load_hash_file( hashfilename, &fsize );
